use float instead of double

+++ mtgoparser/include/mtgoparser/mtgo/card.hpp
@@ -23,8 +23,8 @@ struct [[nodiscard]] Card
   std::string set_;
   mtg::Rarity rarity_;
   bool foil_;
-  double goatbots_price_;
-  std::optional<double> scryfall_price_;
+  float goatbots_price_;
+  std::optional<float> scryfall_price_;
 
 
   // Default constructor
@@ -35,8 +35,8 @@ struct [[nodiscard]] Card
     std::string set = "",
     std::string rarity = "C",
     bool foil = false,
-    double goatbots_price = 0,
-    std::optional<double> scryfall_price = {}) noexcept
+    float goatbots_price = 0,
+    std::optional<float> scryfall_price = {}) noexcept
     : id_{ id }, quantity_{ quantity }, name_{ std::move(name) }, set_{ std::move(set) },
       rarity_{ mtg::util::rarity_from_t(std::move(rarity)) }, foil_{ foil }, goatbots_price_{ goatbots_price },
       scryfall_price_{ scryfall_price }
@@ -49,8 +49,8 @@ struct [[nodiscard]] Card
     const char *set = "",
     const char *rarity = "C",
     bool foil = false,
-    double goatbots_price = 0,
-    std::optional<double> scryfall_price = {}) noexcept
+    float goatbots_price = 0,
+    std::optional<float> scryfall_price = {}) noexcept
     : id_{ id }, quantity_{ quantity }, name_{ name }, set_{ set }, rarity_{ mtg::util::rarity_from_t(rarity) },
       foil_{ foil }, goatbots_price_{ goatbots_price }, scryfall_price_{ scryfall_price }
   {}
@@ -63,23 +63,23 @@ struct [[nodiscard]] Card
     std::string_view set,
     std::string_view rarity,
     bool foil = false,
-    double goatbots_price = 0,
-    std::optional<double> scryfall_price = {}) noexcept
+    float goatbots_price = 0,
+    std::optional<float> scryfall_price = {}) noexcept
     : id_{ id }, quantity_{ quantity }, name_{ name }, set_{ set }, rarity_{ mtg::util::rarity_from_t(rarity) },
       foil_{ foil }, goatbots_price_{ goatbots_price }, scryfall_price_{ scryfall_price }
   {}
 
   // Templated constructor
   template<typename I, typename Q, typename S>
-  requires std::convertible_to<I, uint32_t> && std::convertible_to<Q, uint16_t> && std::convertible_to<S, std::string>
+    requires std::convertible_to<I, uint32_t> && std::convertible_to<Q, uint16_t> && std::convertible_to<S, std::string>
   explicit Card(I id,
     Q quantity,
     S name,
     S set,
     S rarity,
     bool foil = false,
-    double goatbots_price = 0,
-    std::optional<double> scryfall_price = {}) noexcept
+    float goatbots_price = 0,
+    std::optional<float> scryfall_price = {}) noexcept
     : id_{ boost::implicit_cast<uint32_t>(id) }, quantity_{ boost::implicit_cast<uint16_t>(quantity) }, name_{ name },
       set_{ set }, rarity_{ mtg::util::rarity_from_t(rarity) }, foil_{ foil }, goatbots_price_{ goatbots_price },
       scryfall_price_{ scryfall_price }
